+++ pkg/deployment/deployment_inspector.go
@@ -26,6 +26,8 @@ import (
 	"context"
 	"time"
 
+	"github.com/arangodb/kube-arangodb/pkg/apis/deployment"
+
 	api "github.com/arangodb/kube-arangodb/pkg/apis/deployment/v1"
 	"github.com/arangodb/kube-arangodb/pkg/metrics"
 	"github.com/arangodb/kube-arangodb/pkg/util"
@@ -35,10 +37,6 @@ import (
 
 var (
 	inspectDeploymentDurationGauges = metrics.MustRegisterGaugeVec(metricsComponent, "inspect_deployment_duration", "Amount of time taken by a single inspection of a deployment (in sec)", metrics.DeploymentName)
-)
-
-const (
-	arangoDeploymentMaintenanceAnnotation = "deployment.arangodb.com/maintenance"
 )
 
 // inspectDeployment inspects the entire deployment, creates
@@ -74,7 +72,7 @@ func (d *Deployment) inspectDeployment(lastInterval util.Interval) util.Interval
 	} else {
 		// Check if maintenance annotation is set
 		if updated != nil && updated.Annotations != nil {
-			if v, ok := updated.Annotations[arangoDeploymentMaintenanceAnnotation]; ok && v == "true" {
+			if v, ok := updated.Annotations[deployment.ArangoDeploymentPodMaintenanceAnnotation]; ok && v == "true" {
 				// Disable checks if we will enter maintenance mode
 				log.Info().Str("deployment", deploymentName).Msg("Deployment in maintenance mode")
 				return nextInterval