+++ pkg/util/k8sutil/map.go
@@ -22,11 +22,15 @@
 
 package k8sutil
 
-import "regexp"
+import (
+	"regexp"
+
+	"github.com/arangodb/kube-arangodb/pkg/apis/deployment"
+)
 
 const (
 	kubernetesAnnotationMatch = ".*kubernetes\\.io/.*"
-	arangoAnnotationMatch     = ".*arangodb\\.com/"
+	arangoAnnotationMatch     = ".*arangodb\\.com/.*"
 )
 
 var (
@@ -48,6 +52,29 @@ func init() {
 	}
 
 	arangoAnnotationRegex = r
+}
+
+func isFilteredBlockedAnnotation(key string) bool {
+	switch key {
+	case deployment.ArangoDeploymentPodRotateAnnotation:
+		return true
+	default:
+		return false
+	}
+}
+
+func filterBlockedAnnotations(m map[string]string) map[string]string {
+	n := map[string]string{}
+
+	for key, value := range m {
+		if isFilteredBlockedAnnotation(key) {
+			continue
+		}
+
+		n[key] = value
+	}
+
+	return n
 }
 
 // MergeAnnotations into one annotations map
@@ -114,7 +141,12 @@ func filterActualAnnotations(actual, expected map[string]string) map[string]stri
 
 // CompareAnnotations will compare annotations, but will ignore secured annotations which are present in
 // actual but not specified in expected map
+// It will also filter out blocked annotations
 func CompareAnnotations(actual, expected map[string]string) bool {
+	return compareAnnotations(filterBlockedAnnotations(actual), filterBlockedAnnotations(expected))
+}
+
+func compareAnnotations(actual, expected map[string]string) bool {
 	actualFiltered := filterActualAnnotations(actual, expected)
 
 	if actualFiltered == nil && expected == nil {