+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## [master](https://github.com/arangodb/kube-arangodb/tree/master) (N/A)
+- Added annotation to rotate ArangoDeployment in secure way
 
 ## [1.0.0](https://github.com/arangodb/kube-arangodb/tree/1.0.0) (2020-03-03)
 - Removal of v1alpha support for ArangoDeployment, ArangoDeploymentReplication, ArangoBackup